requirements.txt
torch
diffusers
transformers
accelerate
safetensors
pillow
numpy
tqdm